move list-pipes to PipeLibrary
